Moved json data functions to internal/data package
This better structuring suits the project now that we are using Cobra.
